Rename files to use kebab case
